+++ components/bl-simple-modal/src/constants.js
